feat: Implement first subtask humanChoice()
defined variable choice and used prompt function to get player choice which is assigned to choice variable.
Changed the choide of user to lower case and encapsulated these process in infinited loop which is only broken
when the user given value is either rock paper or scissors else it shows alter message and loops again.
Lastly added return to return choice.