feat: Add computerChoice function with random number logic
Implemented the computerChoice function to randomly return one of the
three valid choices: "rock", "paper", or "scissors". This uses
Math.random() and Math.floor() to generate a number from 1 to 3, which
maps to the three possible game option.

This function will be used in each round to simulate the computer's move
in the Rock, Paper, Scissors game.