feat: Implement the final playGame function.
Implemented the playGame function which loops through playRound 5 times , determines the score and display message based on score.

round variable is used to keep trap of iteration and set termination condition.Both humanChoice and computerChoice are called inside loop so that in each iteration there is new value and they are then passed to playRound function.